+++ CHANGELOG.md
@@ -6,7 +6,9 @@ HumHub-RSS-Reader Changelog
 0.1.2 January 25, 2022
 ----------------------
 
-- Feature : Added config option to choose the owner of the RSS items (Issue#4)
+- Feature : config option to choose the owner of the RSS items (Issue#4)
+- Feature : config option to specify maximum age of rss items to import (Issue#9)
+- Feature : config option to specify how far in the future rss feed items can be (Issue#10)
 
 
 0.1.1 January 22, 2022